StandingIndex: double(?) -> int(?)
Nothing reads the magnitude, only compares it to establish order, and no
fixture ever carried a fractional value. cTORS's own TrainGoal already
types it `const int` internally and truncates the wire double at
construction; this makes the type honest about what was already true
downstream. Safe now that every fixture's standingIndex is either a real
integer or null -- no decimal-point literal survives anywhere in the
corpus to trip System.Text.Json's strict int parsing.

Verified end to end: Tests/TestData/scenario_example1.json (real
standingIndex: null entries) deserializes cleanly against the live
Interchange.Scenario type.

+++ ServiceSiteScheduling/Interchange/Scenario.cs
@@ -65,8 +65,10 @@ namespace ServiceSiteScheduling.Interchange
         public IList<IncomingTrainUnit> Members { get; init; } = [];
 
         // The index of the train unit when in- or outstanding, with lower indices
-        // at the A-side of the track
-        public double? StandingIndex { get; set; }
+        // at the A-side of the track. Required and mutually distinct within
+        // inStanding groups sharing a track; see unified-schema-design.md's
+        // "Standing order" decision.
+        public int? StandingIndex { get; set; }
     }
 
     // A request for a train to leave the shunting area
@@ -90,8 +92,9 @@ namespace ServiceSiteScheduling.Interchange
         public IList<TrainUnit>? TrainUnits { get; set; }
 
         // The index of the train unit when in- or outstanding, with lower indices
-        // at the A-side of the track
-        public double? StandingIndex { get; set; }
+        // at the A-side of the track. Optional within outStanding: null means
+        // no preference; see unified-schema-design.md's "Standing order" decision.
+        public int? StandingIndex { get; set; }
     }
 
     // TrainUnit represents a combination of carriages which can move independently.
